mantle/kola: Remove Ignition spec 2 support in tests
Move code shared with new RAID tests to the corresponding file.
